development/playground/requirements.txt
-e ../../libs/violet-storage-lib[parquet]
python-dotenv
pandas
rapidfuzz